LayoutTests/imported/w3c/web-platform-tests/html/canvas/historical.window-expected.txt


PASS CanvasRenderingContext2D.scrollPathIntoView method is removed
